requirements.txt
requests~=2.32.3
pymongo~=4.13.0
matplotlib~=3.10.3
pandas~=2.2.3
python-dotenv~=1.1.0